split code into modules
not necessary but I just wanted to understand simple modules. note that
`pub` is used to denote that a function is public rather than using the
pythonic underscore
